Add Morrowind test plugins
These plugins were hand-made in a hex editor using Dave Humphrey's
documentation for the file format, which is available on UESP.net.

This was done because I don't own a copy of Morrowind, so I can't
use its Construction Set to create plugins using the official tool.
Yes, it was incredibly tedious.